+++ examples/09_config_nice_label.php
@@ -1,0 +1,70 @@
+<?php
+
+/*
+ * Require Performance
+ */
+
+require_once('../vendor/autoload.php');
+use Performance\Performance;
+use Performance\Config;
+
+// Bootstrap class
+$foo = new Foo();
+
+class Foo
+{
+    public function __construct()
+    {
+        // The point label nice wil converter 'CamelCase' to 'Camel case'
+        Config::setPointLabelNice(true);
+
+        $this->synchronizeTaskARun();
+        $this->synchronizeTaskBRun();
+        $this->synchronizeTaskCRun();
+
+        // Finish all tasks and show test results
+        Performance::results();
+    }
+
+    public function synchronizeTaskARun()
+    {
+        // Set point Task A
+        Performance::point(__FUNCTION__);
+
+        //
+        // Run code
+        usleep(2000);
+        //
+
+        // Finish point Task A
+        Performance::finish();
+    }
+
+    public function synchronizeTaskBRun()
+    {
+        // Set point Task B
+        Performance::point(__FUNCTION__);
+
+        //
+        // Run code
+        usleep(2000);
+        //
+
+        // Finish point Task B
+        Performance::finish();
+    }
+
+    public function synchronizeTaskCRun()
+    {
+        // Set point Task C
+        Performance::point(__FUNCTION__);
+
+        //
+        // Run code
+        usleep(2000);
+        //
+
+        // Finish point Task C
+        Performance::finish();
+    }
+}