+++ examples/helper/User.php
@@ -1,0 +1,9 @@
+<?php
+
+class User extends Illuminate\Database\Eloquent\Model
+{
+    protected $table = 'user';
+
+    public $timestamps = false;
+
+}