+++ tests/Unit/T06B_PerformanceMessageTest.php
@@ -1,0 +1,13 @@
+<?php namespace Tests\Unit;
+
+use Performance\Config;
+use Performance\Performance;
+
+class T06B_PerformanceMessageTest extends T06A_PerformanceMessageTest
+{
+    protected function setTestUp()
+    {
+        Performance::instanceReset();
+        Config::setPresenter('web');
+    }
+}