+++ tests/Unit/T08B_ExportTest.php
@@ -1,0 +1,13 @@
+<?php namespace Tests\Unit;
+
+use Performance\Config;
+use Performance\Performance;
+
+class T08B_ExportTest extends T08A_ExportTest
+{
+    protected function setTestUp()
+    {
+        Performance::instanceReset();
+        Config::setPresenter('web');
+    }
+}